Libvirt: fix domain creation for lxc
A condition was introduced during domain creation that needs
disk_info to determine whether or not an instance was booted from
volume. Many calls to _create_domain_and_network were not providing
disk_info and thus the condition was always eventualing to True.

This change adds disk_info to each call that was missing it.

Closes-bug: #1370590

Change-Id: I391de9b8a819c54337b657071451c608fdfef4d6